refactor(web): organize pages into directory-per-page with barrel exports
Move each page (Backup, DataFetcher, Files, Screen, Settings, Update)
into its own directory with index.ts. Rename settings-sections to
settings/sections. Update all import paths.
