+++ web/src/pages/backup/index.ts
@@ -1,0 +1,1 @@
+export { BackupPage } from "./Backup";
