+++ web/src/pages/data-fetcher/index.ts
@@ -1,0 +1,1 @@
+export { DataFetcherPage } from "./DataFetcher";
