+++ web/src/pages/files/index.ts
@@ -1,0 +1,1 @@
+export { FilesPage } from "./Files";
