+++ web/src/pages/screen/index.ts
@@ -1,0 +1,1 @@
+export { ScreenPage } from "./Screen";
