+++ web/src/pages/settings/index.ts
@@ -1,0 +1,1 @@
+export { SettingsPage } from "./Settings";
