+++ web/src/pages/settings/sections/AuthSection.tsx
@@ -1,5 +1,5 @@
-import { useSettings } from "../../context/SettingsContext";
-import { TextField, Card, FormRow } from "../../components/ui";
+import { useSettings } from "../../../context/SettingsContext";
+import { TextField, Card, FormRow } from "../../../components/ui";
 import styles from "./sections.module.css";
 
 export function AuthSection() {
