+++ web/src/pages/update/index.ts
@@ -1,0 +1,1 @@
+export { UpdatePage } from "./Update";
